Codacy: fix last string quotes
I am leaving the console.warn in there, but we
could just as easily remove it? I know logging to
console could be a problem with IE9-10 (if it is
not opened —WTF) but I prefer to include a console
shim in my own code. And there already is another
in the code.