Add sort-option, fix key for selected, and fix double submit (#178)
* Add option to sort select-options

The default option is not-sorted, which is
what it used to be long ago. Not sure what is
preferable.

* Use the key to set the selected value

This fixes #167

* Naive/pragmatic to prevent double form submit

Not sure what caused it, but somehow the form.submit 
was called twice. This prevents double submits 
in a naive but effective way.

* Fix codacy remarks

* Fix codacy remarks, try II :)

* Codacy: fix last string quotes

I am leaving the console.warn in there, but we
could just as easily remove it? I know logging to
console could be a problem with IE9-10 (if it is
not opened —WTF) but I prefer to include a console
shim in my own code. And there already is another
in the code.

* Fixing @NicolasCarpi’s remarks